add: Field descriptions
