--- common/article/__init__.py
@@ -1,4 +1,0 @@
-from ._base import ArticleBase
-from ._article import ArticleCore, Article
-
-__all__ = ["ArticleBase", "ArticleCore", "Article"]
