+++ common/schema/article/__init__.py
@@ -1,0 +1,3 @@
+from ._article import ArticleBase, ArticleCore, Article
+
+__all__ = ["ArticleBase", "ArticleCore", "Article"]
