+++ common/schema/user/__init__.py
@@ -1,0 +1,3 @@
+from ._user import UserBase, User
+
+__all__ = ["UserBase", "User"]
