+++ common/schema/user/_user.py
@@ -3,24 +3,22 @@ import re
 from uuid import UUID
 from datetime import datetime
 
-from pydantic import BaseModel, validator
+from pydantic import BaseModel, validator, Field
+
+
+class UserBase(BaseModel):
+    name: Optional[str]
+    email: Optional[str]
 
 
 class User(BaseModel):
     id: UUID
-    name: str
+    name: str = Field(min_length=1)
     email: str
 
     created_at: datetime
     updated_at: Optional[datetime]
     deleted_at: Optional[datetime]
-
-    # TODO: move it to the router schemas
-    @validator("name")
-    def name_is_not_empty(cls, v):
-        if len(v) == 0:
-            raise ValueError("user name should be non-empty")
-        return v
 
     # TODO: move it to the router schemas
     @validator("email")
