--- common/user/__init__.py
@@ -1,4 +1,0 @@
-from ._base import UserBase
-from ._user import User
-
-__all__ = ["UserBase", "User"]
