--- common/user/_base.py
@@ -1,8 +1,0 @@
-from typing import Optional
-
-from pydantic import BaseModel
-
-
-class UserBase(BaseModel):
-    name: Optional[str] = None
-    email: Optional[str] = None
